+++ src/iosMain/kotlin/uz/yalla/platform/button/NativeSquircleIconButton.ios.kt
@@ -6,12 +6,14 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.graphics.isSpecified
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.viewinterop.UIKitInteropProperties
 import androidx.compose.ui.viewinterop.UIKitViewController
 import kotlinx.cinterop.ExperimentalForeignApi
 import platform.UIKit.UIColor
@@ -21,7 +23,7 @@ import uz.yalla.platform.toAssetName
 
 private val SquircleShape = RoundedCornerShape(12.dp)
 
-@OptIn(ExperimentalForeignApi::class)
+@OptIn(ExperimentalForeignApi::class, ExperimentalComposeUiApi::class)
 @Composable
 actual fun NativeSquircleIconButton(
     iconType: IconType,
@@ -50,7 +52,8 @@ actual fun NativeSquircleIconButton(
                 controller.view.layer.setOpaque(false)
                 controller.view.layer.backgroundColor = UIColor.clearColor.CGColor
             },
-            modifier = Modifier.size(48.dp)
+            modifier = Modifier.size(48.dp),
+            properties = UIKitInteropProperties(placedAsOverlay = true)
         )
     }
 }